+++ plasmapy/diagnostics/thomson.py
@@ -261,7 +261,7 @@ def spectral_density_lite(
     T_i={"can_be_negative": False, "equivalencies": u.temperature_energy()},
 )
 @bind_lite_func(spectral_density_lite)
-def spectral_density(  # noqa: PLR0912, PLR0915
+def spectral_density(  # noqa: C901, PLR0912, PLR0915
     wavelengths: u.nm,
     probe_wavelength: u.nm,
     n: u.m**-3,
@@ -679,7 +679,9 @@ def _spectral_density_model(wavelengths, settings=None, **params):
     return model_Skw
 
 
-def spectral_density_model(wavelengths, settings, params):  # noqa: PLR0912, PLR0915
+def spectral_density_model(  # noqa: C901, PLR0912, PLR0915
+    wavelengths, settings, params
+):
     r"""
     Returns a `lmfit.model.Model` function for Thomson spectral density
     function.
@@ -789,7 +791,7 @@ def spectral_density_model(wavelengths, settings, params):  # noqa: PLR0912, PLR
     # **********************
     for p, nums in zip(["T_e", "T_i"], [num_e, num_i]):
         for num in range(nums):
-            key = f"{p}_{str(num)}"
+            key = f"{p}_{num!s}"
             if key not in params:
                 raise ValueError(
                     f"{p} was not provided in kwarg 'parameters', but is required."