+++ plasmapy/formulary/collisions/tests/test_frequencies.py
@@ -13,8 +13,8 @@ from plasmapy.formulary.collisions.frequencies import (
 )
 from plasmapy.particles import Particle
 from plasmapy.utils import exceptions
+from plasmapy.utils._pytest_helpers import assert_can_handle_nparray
 from plasmapy.utils.exceptions import CouplingWarning, PhysicsError
-from plasmapy.utils.pytest_helpers import assert_can_handle_nparray
 
 
 class TestSingleParticleCollisionFrequencies:
@@ -108,7 +108,7 @@ class TestSingleParticleCollisionFrequencies:
         assert MKS_result == CGS_result
 
     @staticmethod
-    def get_limit_value(interaction_type, limit_type, cases):
+    def get_limit_value(interaction_type, limit_type, cases):  # noqa: C901
         """
         Get the limiting values for frequencies given the two particles
         interacting, and their frequencies class.